🔨️ 파일 이름 변경, 북마크 관련 부분 이동
